#15 - Refactors InstaGramSearchBy* and adds SearchByLocation
