+++ .gitignore
@@ -54,3 +54,5 @@ docs/_build/
 
 # vim swap file
 *.swp
+*~
+tests/*~
